+++ library/Ivoz/Provider/Domain/Model/CompanyRelRoutingTag/CompanyRelRoutingTagInterface.php
@@ -6,20 +6,13 @@ use Ivoz\Core\Domain\Model\LoggableEntityInterface;
 
 interface CompanyRelRoutingTagInterface extends LoggableEntityInterface
 {
+    public function setCompany(\Ivoz\Provider\Domain\Model\Company\CompanyInterface $company = null);
+
     /**
      * @codeCoverageIgnore
      * @return array
      */
     public function getChangeSet();
-
-    /**
-     * Set company
-     *
-     * @param \Ivoz\Provider\Domain\Model\Company\CompanyInterface $company | null
-     *
-     * @return static
-     */
-    public function setCompany(\Ivoz\Provider\Domain\Model\Company\CompanyInterface $company = null);
 
     /**
      * Get company