+++ library/Ivoz/Provider/Domain/Model/CompanyRelRoutingTag/CompanyRelRoutingTagRepository.php
@@ -7,5 +7,9 @@ use Doctrine\Common\Persistence\ObjectRepository;
 
 interface CompanyRelRoutingTagRepository extends ObjectRepository, Selectable
 {
-
+    /**
+     * Used by client API access controls
+     * @return int[]
+     */
+    public function getRoutingTagIdsByCompany(int $companyId): array;
 }