+++ library/Ivoz/Provider/Domain/Model/RatingProfile/RatingProfileRepository.php
@@ -7,5 +7,9 @@ use Doctrine\Common\Persistence\ObjectRepository;
 
 interface RatingProfileRepository extends ObjectRepository, Selectable
 {
-
+    /**
+     * Used by client API access controls
+     * @return int[]
+     */
+    public function getRatingPlanGroupIdsByCompany(int $companyId): array;
 }